@@ -3,7 +3,7 @@ description:
   Flutter plugin for Cloud Firestore, a cloud-hosted, noSQL database with
   live synchronization and offline support on Android and iOS.
 homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/cloud_firestore/cloud_firestore
-version: 0.13.0+2
+version: 0.13.4+2
 
 flutter:
   plugin:
@@ -13,15 +13,26 @@ flutter:
         pluginClass: CloudFirestorePlugin
       ios:
         pluginClass: FLTCloudFirestorePlugin
+      macos:
+        pluginClass: FLTCloudFirestorePlugin
+      web:
+        default_package: cloud_firestore_web
 
 dependencies:
   flutter:
     sdk: flutter
-  meta: "^1.0.5"
-  collection: "^1.14.3"
-  firebase_core: "^0.4.0"
+  meta: ^1.0.5
+  firebase_core: ^0.4.4
+  cloud_firestore_platform_interface: ^1.0.0
+  # The design on https://flutter.dev/go/federated-plugins was to leave
+  # this constraint as "any". We cannot do it right now as it fails pub publish
+  # validation, so we set a ^ constraint.
+  # TODO(amirh): Revisit this (either update this part in the  design or the pub tool).
+  # https://github.com/flutter/flutter/issues/46264
+  cloud_firestore_web: ^0.1.1
 
 dev_dependencies:
+  pedantic: ^1.8.0
   flutter_test:
     sdk: flutter
   flutter_driver:
@@ -30,4 +41,4 @@ dev_dependencies:
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  flutter: ">=1.12.13+hotfix.4 <2.0.0"
