@@ -6,10 +6,14 @@ package io.flutter.plugins.firebaseadmob;
 
 import android.app.Activity;
 import android.content.Context;
+import android.util.Log;
 import android.view.Gravity;
 import com.google.android.gms.ads.AdSize;
 import com.google.android.gms.ads.MobileAds;
+import com.google.android.gms.ads.formats.UnifiedNativeAd;
+import com.google.android.gms.ads.formats.UnifiedNativeAdView;
 import com.google.firebase.FirebaseApp;
+import io.flutter.embedding.engine.FlutterEngine;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
@@ -18,7 +22,9 @@ import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
+import io.flutter.plugin.common.PluginRegistry;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
+import java.util.HashMap;
 import java.util.Locale;
 import java.util.Map;
 
@@ -28,12 +34,39 @@ import java.util.Map;
  * <p>Instantiate this in an add to app scenario to gracefully handle activity and context changes.
  */
 public class FirebaseAdMobPlugin implements FlutterPlugin, ActivityAware, MethodCallHandler {
+  // This the plugin key used in the Embedding V1 generated GeneratedRegistrant.java. This key is
+  // used when this plugin publishes it's self in registerWith(registrar).
+  private static final String GENERATED_PLUGIN_KEY =
+      "io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin";
+
   private Context applicationContext;
   private MethodChannel channel;
   private Activity activity;
   // This is always null when not using v2 embedding.
   private FlutterPluginBinding pluginBinding;
   private RewardedVideoAdWrapper rewardedWrapper;
+  private final Map<String, NativeAdFactory> nativeAdFactories = new HashMap<>();
+
+  /**
+   * Interface used to display a {@link com.google.android.gms.ads.formats.UnifiedNativeAd}.
+   *
+   * <p>Added to a {@link io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin} and creates {@link
+   * com.google.android.gms.ads.formats.UnifiedNativeAdView}s from Native Ads created in Dart.
+   */
+  public interface NativeAdFactory {
+    /**
+     * Creates a {@link com.google.android.gms.ads.formats.UnifiedNativeAdView} with a {@link
+     * com.google.android.gms.ads.formats.UnifiedNativeAd}.
+     *
+     * @param nativeAd Ad information used to create a {@link
+     *     com.google.android.gms.ads.formats.UnifiedNativeAdView}
+     * @param customOptions Used to pass additional custom options to create the {@link
+     *     com.google.android.gms.ads.formats.UnifiedNativeAdView}. Nullable.
+     * @return a {@link com.google.android.gms.ads.formats.UnifiedNativeAdView} that is overlaid on
+     *     top of the FlutterView.
+     */
+    UnifiedNativeAdView createNativeAd(UnifiedNativeAd nativeAd, Map<String, Object> customOptions);
+  }
 
   /**
    * Registers a plugin with the v1 embedding api {@code io.flutter.plugin.common}.
@@ -53,7 +86,116 @@ public class FirebaseAdMobPlugin implements FlutterPlugin, ActivityAware, Method
     }
 
     final FirebaseAdMobPlugin plugin = new FirebaseAdMobPlugin();
+    registrar.publish(plugin);
     plugin.initializePlugin(registrar.context(), registrar.activity(), registrar.messenger());
+  }
+
+  /**
+   * Adds a {@link io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin.NativeAdFactory} used to
+   * create {@link com.google.android.gms.ads.formats.UnifiedNativeAdView}s from a Native Ad created
+   * in Dart.
+   *
+   * @param registry maintains access to a FirebaseAdMobPlugin instance.
+   * @param factoryId a unique identifier for the ad factory. The Native Ad created in Dart includes
+   *     a parameter that refers to this.
+   * @param nativeAdFactory creates {@link com.google.android.gms.ads.formats.UnifiedNativeAdView}s
+   *     when Flutter NativeAds are created.
+   * @return whether the factoryId is unique and the nativeAdFactory was successfully added.
+   */
+  public static boolean registerNativeAdFactory(
+      PluginRegistry registry, String factoryId, NativeAdFactory nativeAdFactory) {
+    final FirebaseAdMobPlugin adMobPlugin = registry.valuePublishedByPlugin(GENERATED_PLUGIN_KEY);
+    return registerNativeAdFactory(adMobPlugin, factoryId, nativeAdFactory);
+  }
+
+  /**
+   * Registers a {@link io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin.NativeAdFactory} used
+   * to create {@link com.google.android.gms.ads.formats.UnifiedNativeAdView}s from a Native Ad
+   * created in Dart.
+   *
+   * @param engine maintains access to a FirebaseAdMobPlugin instance.
+   * @param factoryId a unique identifier for the ad factory. The Native Ad created in Dart includes
+   *     a parameter that refers to this.
+   * @param nativeAdFactory creates {@link com.google.android.gms.ads.formats.UnifiedNativeAdView}s
+   *     when Flutter NativeAds are created.
+   * @return whether the factoryId is unique and the nativeAdFactory was successfully added.
+   */
+  public static boolean registerNativeAdFactory(
+      FlutterEngine engine, String factoryId, NativeAdFactory nativeAdFactory) {
+    final FirebaseAdMobPlugin adMobPlugin =
+        (FirebaseAdMobPlugin) engine.getPlugins().get(FirebaseAdMobPlugin.class);
+    return registerNativeAdFactory(adMobPlugin, factoryId, nativeAdFactory);
+  }
+
+  private static boolean registerNativeAdFactory(
+      FirebaseAdMobPlugin plugin, String factoryId, NativeAdFactory nativeAdFactory) {
+    if (plugin == null) {
+      final String message =
+          String.format(
+              "Could not find a %s instance. The plugin may have not been registered.",
+              FirebaseAdMobPlugin.class.getSimpleName());
+      throw new IllegalStateException(message);
+    }
+
+    return plugin.addNativeAdFactory(factoryId, nativeAdFactory);
+  }
+
+  /**
+   * Unregisters a {@link io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin.NativeAdFactory} used
+   * to create {@link com.google.android.gms.ads.formats.UnifiedNativeAdView}s from a Native Ad
+   * created in Dart.
+   *
+   * @param registry maintains access to a FirebaseAdMobPlugin instance.
+   * @param factoryId a unique identifier for the ad factory. The Native ad created in Dart includes
+   *     a parameter that refers to this.
+   * @return the previous {@link
+   *     io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin.NativeAdFactory} associated with this
+   *     factoryId, or null if there was none for this factoryId.
+   */
+  public static NativeAdFactory unregisterNativeAdFactory(
+      PluginRegistry registry, String factoryId) {
+    final FirebaseAdMobPlugin adMobPlugin = registry.valuePublishedByPlugin(GENERATED_PLUGIN_KEY);
+    if (adMobPlugin != null) adMobPlugin.removeNativeAdFactory(factoryId);
+
+    return null;
+  }
+
+  /**
+   * Unregisters a {@link io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin.NativeAdFactory} used
+   * to create {@link com.google.android.gms.ads.formats.UnifiedNativeAdView}s from a Native Ad
+   * created in Dart.
+   *
+   * @param engine maintains access to a FirebaseAdMobPlugin instance.
+   * @param factoryId a unique identifier for the ad factory. The Native ad created in Dart includes
+   *     a parameter that refers to this.
+   * @return the previous {@link
+   *     io.flutter.plugins.firebaseadmob.FirebaseAdMobPlugin.NativeAdFactory} associated with this
+   *     factoryId, or null if there was none for this factoryId.
+   */
+  public static NativeAdFactory unregisterNativeAdFactory(FlutterEngine engine, String factoryId) {
+    final FlutterPlugin adMobPlugin = engine.getPlugins().get(FirebaseAdMobPlugin.class);
+    if (adMobPlugin != null) {
+      return ((FirebaseAdMobPlugin) adMobPlugin).removeNativeAdFactory(factoryId);
+    }
+
+    return null;
+  }
+
+  private boolean addNativeAdFactory(String factoryId, NativeAdFactory nativeAdFactory) {
+    if (nativeAdFactories.containsKey(factoryId)) {
+      final String errorMessage =
+          String.format(
+              "A NativeAdFactory with the following factoryId already exists: %s", factoryId);
+      Log.e(FirebaseAdMobPlugin.class.getSimpleName(), errorMessage);
+      return false;
+    }
+
+    nativeAdFactories.put(factoryId, nativeAdFactory);
+    return true;
+  }
+
+  private NativeAdFactory removeNativeAdFactory(String factoryId) {
+    return nativeAdFactories.remove(factoryId);
   }
 
   private void initializePlugin(
@@ -75,6 +217,41 @@ public class FirebaseAdMobPlugin implements FlutterPlugin, ActivityAware, Method
       return;
     }
     MobileAds.initialize(applicationContext, appId);
+    result.success(Boolean.TRUE);
+  }
+
+  private void callLoadNativeAd(Integer id, Activity activity, MethodCall call, Result result) {
+    final String adUnitId = call.argument("adUnitId");
+    final String factoryId = call.argument("factoryId");
+
+    final NativeAdFactory nativeAdFactory = nativeAdFactories.get(factoryId);
+
+    if (adUnitId == null || adUnitId.isEmpty()) {
+      result.error("no_unit_id", "a null or empty adUnitId was provided for ad id=" + id, null);
+      return;
+    } else if (nativeAdFactory == null) {
+      final String errorMessage =
+          String.format(
+              "There is no non-null %s for the following factoryId: %s",
+              NativeAdFactory.class.getSimpleName(), factoryId);
+      result.error("no_native_ad_factory", errorMessage, null);
+      return;
+    }
+
+    final Map<String, Object> customOptions = call.argument("customOptions");
+
+    final MobileAd.Native nativeAd =
+        MobileAd.createNative(id, activity, channel, nativeAdFactory, customOptions);
+
+    if (nativeAd.status != MobileAd.Status.CREATED) {
+      if (nativeAd.status == MobileAd.Status.FAILED)
+        result.error("load_failed_ad", "cannot reload a failed ad, id=" + id, null);
+      else result.success(Boolean.TRUE); // The ad was already loaded.
+      return;
+    }
+
+    final Map<String, Object> targetingInfo = call.argument("targetingInfo");
+    nativeAd.load(adUnitId, targetingInfo);
     result.success(Boolean.TRUE);
   }
 
@@ -217,6 +394,20 @@ public class FirebaseAdMobPlugin implements FlutterPlugin, ActivityAware, Method
     }
   }
 
+  private void callSetRewardedVideoAdUserId(MethodCall call, Result result) {
+    String userId = call.argument("userId");
+
+    rewardedWrapper.setUserId(userId);
+    result.success(Boolean.TRUE);
+  }
+
+  private void callSetRewardedVideoAdCustomData(MethodCall call, Result result) {
+    String customData = call.argument("customData");
+
+    rewardedWrapper.setCustomData(customData);
+    result.success(Boolean.TRUE);
+  }
+
   private void callDisposeAd(Integer id, Result result) {
     MobileAd ad = MobileAd.getAdForId(id);
     if (ad == null) {
@@ -243,7 +434,7 @@ public class FirebaseAdMobPlugin implements FlutterPlugin, ActivityAware, Method
     initializePlugin(
         pluginBinding.getApplicationContext(),
         binding.getActivity(),
-        pluginBinding.getFlutterEngine().getDartExecutor());
+        pluginBinding.getBinaryMessenger());
   }
 
   @Override
@@ -257,7 +448,7 @@ public class FirebaseAdMobPlugin implements FlutterPlugin, ActivityAware, Method
     initializePlugin(
         pluginBinding.getApplicationContext(),
         binding.getActivity(),
-        pluginBinding.getFlutterEngine().getDartExecutor());
+        pluginBinding.getBinaryMessenger());
   }
 
   @Override
@@ -288,11 +479,20 @@ public class FirebaseAdMobPlugin implements FlutterPlugin, ActivityAware, Method
       case "loadRewardedVideoAd":
         callLoadRewardedVideoAd(call, result);
         break;
+      case "loadNativeAd":
+        callLoadNativeAd(id, activity, call, result);
+        break;
       case "showAd":
         callShowAd(id, call, result);
         break;
       case "showRewardedVideoAd":
         callShowRewardedVideoAd(result);
+        break;
+      case "setRewardedVideoAdUserId":
+        callSetRewardedVideoAdUserId(call, result);
+        break;
+      case "setRewardedVideoAdCustomData":
+        callSetRewardedVideoAdCustomData(call, result);
         break;
       case "disposeAd":
         callDisposeAd(id, result);
