@@ -3,6 +3,7 @@
 // found in the LICENSE file.
 
 #import <Flutter/Flutter.h>
+#import "FLTFirebaseAdMobPlugin.h"
 #import "GoogleMobileAds/GoogleMobileAds.h"
 
 typedef enum : NSUInteger {
@@ -25,7 +26,11 @@ typedef enum : NSUInteger {
 - (void)dispose;
 @end
 
-@interface FLTBannerAd : FLTMobileAd <GADBannerViewDelegate>
+@interface FLTMobileAdWithView : FLTMobileAd
+- (UIView *)adView;
+@end
+
+@interface FLTBannerAd : FLTMobileAdWithView <GADBannerViewDelegate>
 + (instancetype)withId:(NSNumber *)mobileAdId
                 adSize:(GADAdSize)adSize
                channel:(FlutterMethodChannel *)channel;
@@ -33,4 +38,12 @@ typedef enum : NSUInteger {
 
 @interface FLTInterstitialAd : FLTMobileAd <GADInterstitialDelegate>
 + (instancetype)withId:(NSNumber *)mobileAdId channel:(FlutterMethodChannel *)channel;
+@end
+
+@interface FLTNativeAd
+    : FLTMobileAdWithView <GADUnifiedNativeAdLoaderDelegate, GADUnifiedNativeAdDelegate>
++ (instancetype)withId:(NSNumber *)mobileAdId
+               channel:(FlutterMethodChannel *)channel
+       nativeAdFactory:(id<FLTNativeAdFactory>)nativeAdFactory
+         customOptions:(NSDictionary *)customOptions;
 @end
