@@ -83,7 +83,7 @@ int _anchorType;
 }
 
 - (void)show {
-  // Implemented by the Banner and Interstitial subclasses
+  // Implemented by the FLTMobileAdWithView and Interstitial subclasses
 }
 
 - (void)dispose {
@@ -98,6 +98,74 @@ int _anchorType;
   NSString *statusString = (NSString *)statusToString[[NSNumber numberWithInt:_status]];
   return [NSString
       stringWithFormat:@"%@ %@ mobileAdId:%@", super.description, statusString, _mobileAdId];
+}
+@end
+
+@implementation FLTMobileAdWithView
+- (UIView *)adView {
+  // We cause a crash if this method is not overriden by subclasses.
+  [self doesNotRecognizeSelector:_cmd];
+  return nil;
+}
+
+- (void)show {
+  if (_status == LOADING) {
+    _status = PENDING;
+    return;
+  }
+
+  if (_status != LOADED) return;
+
+  UIView *screen = [FLTMobileAd rootViewController].view;
+  [screen addSubview:self.adView];
+
+// UIView.safeAreaLayoutGuide is only available on iOS 11.0+
+#if defined(__IPHONE_11_0) && (__IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_11_0)
+  if (@available(ios 11.0, *)) {
+    self.adView.translatesAutoresizingMaskIntoConstraints = NO;
+
+    UILayoutGuide *guide = screen.safeAreaLayoutGuide;
+    [NSLayoutConstraint activateConstraints:@[
+      [self.adView.centerXAnchor constraintEqualToAnchor:guide.centerXAnchor
+                                                constant:_horizontalCenterOffset],
+
+      [self.adView.leftAnchor constraintGreaterThanOrEqualToAnchor:guide.leftAnchor],
+      [self.adView.rightAnchor constraintLessThanOrEqualToAnchor:guide.rightAnchor],
+    ]];
+
+    if (_anchorType == 0) {
+      [NSLayoutConstraint activateConstraints:@[
+        [self.adView.bottomAnchor constraintEqualToAnchor:guide.bottomAnchor
+                                                 constant:_anchorOffset],
+      ]];
+    } else {
+      [NSLayoutConstraint activateConstraints:@[
+        [self.adView.topAnchor constraintEqualToAnchor:guide.topAnchor constant:_anchorOffset],
+      ]];
+    }
+  }
+#endif
+
+  // We find the left most point that aligns the view to the horizontal center.
+  CGFloat x =
+      screen.frame.size.width / 2 - self.adView.frame.size.width / 2 + _horizontalCenterOffset;
+  // We find the top point that anchors the view to the top/bottom depending on anchorType.
+  CGFloat y;
+  if (_anchorType == 0) {
+    y = screen.frame.size.height - self.adView.frame.size.height + _anchorOffset;
+  } else {
+    y = _anchorOffset;
+  }
+  self.adView.frame = (CGRect){{x, self.adView.frame.origin.y}, self.adView.frame.size};
+}
+
+- (void)dispose {
+  if (self.adView.superview) [self.adView removeFromSuperview];
+  [super dispose];
+}
+
+- (NSString *)description {
+  return [NSString stringWithFormat:@"%@ for: %@", super.description, self.adView];
 }
 @end
 
@@ -125,6 +193,10 @@ GADAdSize _adSize;
   return nil;
 }
 
+- (UIView *)adView {
+  return _banner;
+}
+
 - (void)loadWithAdUnitId:(NSString *)adUnitId targetingInfo:(NSDictionary *)targetingInfo {
   if (_status != CREATED) return;
   _status = LOADING;
@@ -134,56 +206,6 @@ GADAdSize _adSize;
   _banner.rootViewController = [FLTMobileAd rootViewController];
   FLTRequestFactory *factory = [[FLTRequestFactory alloc] initWithTargetingInfo:targetingInfo];
   [_banner loadRequest:[factory createRequest]];
-}
-
-- (void)show {
-  if (_status == LOADING) {
-    _status = PENDING;
-    return;
-  }
-
-  if (_status != LOADED) return;
-
-  _banner.translatesAutoresizingMaskIntoConstraints = NO;
-  UIView *screen = [FLTMobileAd rootViewController].view;
-  [screen addSubview:_banner];
-
-#if defined(__IPHONE_11_0) && (__IPHONE_OS_VERSION_MAX_ALLOWED >= __IPHONE_11_0)
-  if (@available(ios 11.0, *)) {
-    UILayoutGuide *guide = screen.safeAreaLayoutGuide;
-    [NSLayoutConstraint activateConstraints:@[
-      [_banner.centerXAnchor constraintEqualToAnchor:guide.centerXAnchor
-                                            constant:_horizontalCenterOffset],
-      [_banner.bottomAnchor
-          constraintEqualToAnchor:_anchorType == 0 ? guide.bottomAnchor : guide.topAnchor
-                         constant:_anchorOffset]
-    ]];
-  } else {
-    [self placeBannerPreIos11];
-  }
-#else
-  [self placeBannerPreIos11];
-#endif
-}
-
-- (void)placeBannerPreIos11 {
-  UIView *screen = [FLTMobileAd rootViewController].view;
-  CGFloat x = screen.frame.size.width / 2 - _banner.frame.size.width / 2 + _horizontalCenterOffset;
-  CGFloat y;
-  if (_anchorType == 0) {
-    y = screen.frame.size.height - _banner.frame.size.height + _anchorOffset;
-  } else {
-    y = _anchorOffset;
-  }
-  _banner.frame = (CGRect){{x, y}, _banner.frame.size};
-  [screen addSubview:_banner];
-}
-
-- (void)adViewDidReceiveAd:(GADBannerView *)adView {
-  bool statusWasPending = _status == PENDING;
-  _status = LOADED;
-  [_channel invokeMethod:@"onAdLoaded" arguments:[self argumentsMap]];
-  if (statusWasPending) [self show];
 }
 
 - (void)adView:(GADBannerView *)adView didFailToReceiveAdWithError:(GADRequestError *)error {
@@ -208,14 +230,11 @@ GADAdSize _adSize;
   [_channel invokeMethod:@"onAdLeftApplication" arguments:[self argumentsMap]];
 }
 
-- (void)dispose {
-  if (_banner.superview) [_banner removeFromSuperview];
-  _banner = nil;
-  [super dispose];
-}
-
-- (NSString *)description {
-  return [NSString stringWithFormat:@"%@ for: %@", super.description, _banner];
+- (void)adViewDidReceiveAd:(GADBannerView *)adView {
+  bool statusWasPending = _status == PENDING;
+  _status = LOADED;
+  [_channel invokeMethod:@"onAdLoaded" arguments:[self argumentsMap]];
+  if (statusWasPending) [self show];
 }
 @end
 
@@ -285,5 +304,89 @@ GADInterstitial *_interstitial;
 
 - (NSString *)description {
   return [NSString stringWithFormat:@"%@ for: %@", super.description, _interstitial];
+}
+@end
+
+@implementation FLTNativeAd {
+  GADAdLoader *_adLoader;
+  GADUnifiedNativeAdView *_nativeAd;
+  id<FLTNativeAdFactory> _nativeAdFactory;
+  NSDictionary *_customOptions;
+}
+
++ (instancetype)withId:(NSNumber *)mobileAdId
+               channel:(FlutterMethodChannel *)channel
+       nativeAdFactory:(id<FLTNativeAdFactory>)nativeAdFactory
+         customOptions:(NSDictionary *)customOptions {
+  FLTMobileAd *ad = [FLTMobileAd getAdForId:mobileAdId];
+  return ad != nil ? (FLTNativeAd *)ad
+                   : [[FLTNativeAd alloc] initWithId:mobileAdId
+                                             channel:channel
+                                     nativeAdFactory:nativeAdFactory
+                                       customOptions:customOptions];
+}
+
+- (instancetype)initWithId:mobileAdId
+                   channel:(FlutterMethodChannel *)channel
+           nativeAdFactory:(id<FLTNativeAdFactory>)nativeAdFactory
+             customOptions:(NSDictionary *)customOptions {
+  self = [super initWithId:mobileAdId channel:channel];
+  if (self) {
+    _nativeAdFactory = nativeAdFactory;
+    _customOptions = customOptions;
+  }
+  return self;
+}
+
+- (UIView *)adView {
+  return _nativeAd;
+}
+
+- (void)loadWithAdUnitId:(NSString *)adUnitId targetingInfo:(NSDictionary *)targetingInfo {
+  if (_status != CREATED) return;
+  _status = LOADING;
+
+  _adLoader = [[GADAdLoader alloc] initWithAdUnitID:adUnitId
+                                 rootViewController:[FLTMobileAd rootViewController]
+                                            adTypes:@[ kGADAdLoaderAdTypeUnifiedNative ]
+                                            options:@[]];
+  _adLoader.delegate = self;
+
+  FLTRequestFactory *factory = [[FLTRequestFactory alloc] initWithTargetingInfo:targetingInfo];
+  [_adLoader loadRequest:[factory createRequest]];
+}
+
+- (void)adLoader:(nonnull GADAdLoader *)adLoader
+    didFailToReceiveAdWithError:(nonnull GADRequestError *)error {
+  FLTLogWarning(@"adLoader:didFailToReceiveAdWithError: %@ (MobileAd %@)",
+                [error localizedDescription], self);
+  [_channel invokeMethod:@"onAdFailedToLoad" arguments:[self argumentsMap]];
+}
+
+- (void)adLoader:(nonnull GADAdLoader *)adLoader
+    didReceiveUnifiedNativeAd:(nonnull GADUnifiedNativeAd *)nativeAd {
+  nativeAd.delegate = self;
+  _nativeAd = [_nativeAdFactory createNativeAd:nativeAd customOptions:_customOptions];
+
+  bool statusWasPending = _status == PENDING;
+  _status = LOADED;
+  [_channel invokeMethod:@"onAdLoaded" arguments:[self argumentsMap]];
+  if (statusWasPending) [self show];
+}
+
+- (void)nativeAdWillPresentScreen:(GADUnifiedNativeAd *)nativeAd {
+  [_channel invokeMethod:@"onAdClicked" arguments:[self argumentsMap]];
+}
+
+- (void)nativeAdWillDismissScreen:(GADUnifiedNativeAd *)nativeAd {
+  [_channel invokeMethod:@"onAdImpression" arguments:[self argumentsMap]];
+}
+
+- (void)nativeAdDidDismissScreen:(GADUnifiedNativeAd *)nativeAd {
+  [_channel invokeMethod:@"onAdClosed" arguments:[self argumentsMap]];
+}
+
+- (void)nativeAdWillLeaveApplication:(GADUnifiedNativeAd *)nativeAd {
+  [_channel invokeMethod:@"onAdLeftApplication" arguments:[self argumentsMap]];
 }
 @end
