@@ -8,6 +8,7 @@ import 'dart:async';
 
 import 'package:flutter/services.dart';
 import 'package:meta/meta.dart' show required, visibleForTesting;
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 part 'src/method_channel_firebase_auth.dart';
 part 'src/types.dart';
@@ -20,14 +21,10 @@ part 'src/types.dart';
 /// will get the default implementation, while platform implementations that
 /// `implements` this interface will be broken by newly added
 /// [FirebaseAuthPlatform] methods.
-abstract class FirebaseAuthPlatform {
-  /// Only mock implementations should set this to `true`.
-  ///
-  /// Mockito mocks implement this class with `implements` which is forbidden
-  /// (see class docs). This property provides a backdoor for mocks to skip the
-  /// verification that the class isn't implemented with `implements`.
-  @visibleForTesting
-  bool get isMock => false;
+abstract class FirebaseAuthPlatform extends PlatformInterface {
+  FirebaseAuthPlatform() : super(token: _token);
+
+  static final Object _token = Object();
 
   /// The default instance of [FirebaseAuthPlatform] to use.
   ///
@@ -39,28 +36,12 @@ abstract class FirebaseAuthPlatform {
 
   static FirebaseAuthPlatform _instance = MethodChannelFirebaseAuth();
 
-  // TODO(amirh): Extract common platform interface logic.
-  // https://github.com/flutter/flutter/issues/43368
+  /// Platform-specific plugins should set this with their own platform-specific
+  /// class that extends [FirebaseAuthPlatform] when they register themselves.
   static set instance(FirebaseAuthPlatform instance) {
-    if (!instance.isMock) {
-      try {
-        instance._verifyProvidesDefaultImplementations();
-      } on NoSuchMethodError catch (_) {
-        throw AssertionError(
-            'Platform interfaces must not be implemented with `implements`');
-      }
-    }
+    PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }
-
-  /// This method ensures that [FirebaseAuthPlatform] isn't implemented with `implements`.
-  ///
-  /// See class docs for more details on why using `implements` to implement
-  /// [FirebaseAuthPlatform] is forbidden.
-  ///
-  /// This private method is called by the [instance] setter, which should fail
-  /// if the provided instance is a class implemented with `implements`.
-  void _verifyProvidesDefaultImplementations() {}
 
   /// Returns the current user.
   Future<PlatformUser> getCurrentUser(String app) {
